requirements.txt
# Core Dependencies
python-daemon>=3.0.0
psutil>=5.9.0
watchdog>=3.0.0

# Database
chromadb>=0.4.0
sentence-transformers>=2.2.0
networkx>=3.1

# API & Web
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0

# Monitoring
pyxhook3>=1.0.0  # Keyboard/mouse monitoring
python-xlib>=0.33  # X11 window tracking
pydbus>=0.6.0  # D-Bus for system integration

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
schedule>=1.2.0
